Moved to a syntax package and implicit conversions in its package object
